refactor: Move users router into auth module
